Improve basic integration test
